viced: time_t might not be long
Fix a couple of printf format errors that bite on FreeBSD 10 for i386.
Since time_t might be an int, it can't be printed with a long format.
Since time_t might be a long in general, cast to it to long when
printing.

Reviewed-on: http://gerrit.openafs.org/11385
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry picked from commit 9dd67783bb2bd9ef8a972a1aac47b1925069a655)

Reviewed-on: http://gerrit.openafs.org/11404
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ src/viced/callback.c
@@ -2961,7 +2961,7 @@ PrintCB(struct CallBack *cb, afs_uint32 now)
 
     printf("vol=%u vn=%u cbs=%d hi=%d st=%d fest=%d, exp in %lu secs at %s",
 	   fe->volid, fe->vnode, fe->ncbs, cb->hhead, cb->status, fe->status,
-	   expires - now, ctime(&expires));
+	   (unsigned long)(expires - now), ctime(&expires));
 }
 
 #endif
